Support for character literals
This adds support for all three forms of character literals:

- single-character literals like #\!
- named character literals like #\newline
- hexcoded Unicode literals like #\x200C

As noted in comments, such syntax is unbelievably ambiguous and
generally awful, but we have to deal with it because of raisins.
After all, string manipulation is not a design goal for Scheme.

Though, there are some points which I could not stand and opted
to digress from R7RS in order to implement a saner behavior.

For example, R7RS demands the following parses:

    "#\\ "    -> Character(' ')
    "#\\\t"   -> Character('\t')
    "#\\\n"   -> Character('\n')
    "#\\\r\n" -> unspecified lol,
                 supposedly Character('\r') + Whitespace

I find this extremely confusing and I believe that whitespace
should not matter outside of string literals and escaped
identifiers. Thus in this case we will produce a syntax error
and will return a placeholder character instead. The following
whitespace will be ignored then. We insist on using named
characters like #\space in these cases.

Also, delimiter rules have been tweaked a bit and applied to the
unrecogized sequence parser, so recovery test cases for vector
openers had to be updated.