Support for string literals
This adds support for string literals and all escape sequences
permitted in them. This commit also adds the _intern pool_ to
the library as strings are the first tokens without immediate
representation. The pool will be used to store symbolic data
output by the scanner such as string values, identifier names,
exact number spelling, directive names, etc. The pool has some
details regarding its usage, refer to module documentation.

Support for all escape sequences is kinda bulky, especially for
the line escapes. But again, this is the same story as with
character literals: this is legacy and we have to deal with it.
I also like proper error recovery so we do that as well.