Support for integer numbers
This adds support for integer number forms of Scheme, including plain
decimal numbers, explicit signs, and base and exactness prefixes.

As with characters, the syntax is terribad, but this is even worse as
numbers are competing with identifiers which have very liberal syntax.
The disambiguation rule is literally if the prefix is a number then it
is a number, otherwise it is an identifier. But again, this is life.

Number prefixes start with a hash, so this change also updates recovery
in (byte)vector literals. These are not significant changes though.
The main point is that we now use self.peek() in scan_hash_token()
in order to leave self.cur == '#' when we call scan_number_prefixed().

Also, note that we do not parse numbers here (i.e., they all are flat
Token::Number, not Token::Number::Integer::DecimalImplicitPositive).
This is because the parser does not care for exact number types and
their values, and we will still have exact spelling of the tokens
when we will need it.