Support for float numbers
This adds support for float number forms of Scheme, including decimal
point floats, exponential (scientific) notation, and IEEE 754 special
values.

Again, terrible syntax and terrible workarounds, but this is also a
terrible implementation. I will be refactoring it in some next commit.
Now it is important to have some code that works.