Do not treat # as a delimiter
After all, R7RS says that # is *not* a delimiter. And we do not treat it
as such in number scanning code. Just quit pretending that it is a good
idea and do not treat # as a delimiter anywhere.

This makes chaining characters like #\t#\e#\s#\t invalid, update the
tests to reflect this fact.

One important note on this is that other separators are fine, so
things like #\t|he|,#\s"e" #\a(re) #\o;k.

Also introduce a helper function to check for whitespace
(scan_character_literal() looked ugly without it).