Improve handling of infnans in numbers
Basically we are moving handling of explicit signs and infnans deeper
into the number scanning code. This simplifies their handling in the
case of rational numbers, and it should make implementation of complex
numbers easier.

For example, this allows us to detect and report infnans in denominators
instead of treating the whole rational number as an invalid suffix of
the initial infnan, or considering them simply invalid characters.

However, the tests are still broken due to bad handling of peculiar
identifiers (which cases like "+infinity" are). This is the evil part
of Scheme's grammar. I'm not responsible for this insanity that demands
"+inf.0/inf.0" to be treated as an invalid number, but "+inf.O/inf.O"
being a totally valid identifier. But, meh, legacy. I will make a rant
about it in comments later.

(Yes, infnan handling now requires 5-character lookahead, but it makes
the implementation so much easier and readble that I have decided to
leave it as is. Blame the peculiar identifiers and their syntax.)