Improve handling of peculiar identifiers
Yes, they are awful.

This improves handling of special identifiers like +, ..., and so on. All
of them start like a number while not being one in the end. To simplify
parsing of actual numbers we check for peculiar identifiers at start and
do not return to this question again. This allows us to assume that we are
scanning a valid number and anything invalid is an error, not an attempt
to write a peculiar identifier.

Except for special cases like '+inf.0', there is another special case for
recovery: a peculiar identifier followed by a number prefix. "#infinity",
for example. Previously we have been doing a quick check for numbers, but
now we do a quick check for peculiar identifiers in scan_number_literal(),
so we can just skip to there from the hash prefix scanning code.

The actual code for checking is a bit convoluted. Hopefully, the formal
grammar in the comments will be helpful in understanding it.

I am not particularly happy with the diagnostics produced by this
recovery, but they are okayish, considering the circumstances. It is
much better that spewing out an err_lexer_invalid_number_character
for every character of a token.