Treat +. as peculiar identifier
This is to be consistent with treatment of ".+". The grammar of Scheme
does not explicitly specify how to treat this sequence of characters
(presumably, it is invalid). However, we treat it as a peculiar identifier
of form <explicit-sign> <dot> <dot-subsequent> (class 3 as recognized
by peculiar_identifier_ahead).

Signs are pretty important delimiters in complex numbers, so getting this
straight is also important.