Support for polar form of complex numbers
This is actually the sanest form of complex numbers because it has
a distinct sigil in it which is not used anywhere else. Rectangular
form heavily uses signs so it is much harder to implement *and*
(mostly) to recover from errors.

Scanning of special characters will be controlled by the Complex-
ScanningMode enum. We only add success tests there as there isn't
much new recovery now.