Support for rectangular form of complex numbers
Now, this is hard. Mostly because the rectangular form uses + and - signs
as delimiters, and they are already used in multiple places in numbers.
Plus it has wacky special cases like '+i' and '-i', and infnans with their
suffixes require special treatment as well, and many more stuff.

However, this _is_ mostly the right thing. Though, the tests are currently
broken because of several deficiencies in handling sign characters. These
will be rectified in the next commits.

There are some changes in previous non-complex tests. They are related to
the fact that sign characters now mark the start of an imaginary part of
a complex number. So in these places they are not invalid characters
anymore, and we expect an 'i' to appear in the end.

This also adds recovery tests for complex numbers (including polar form).
Note how some of the strings are parsed as peculiar identifiers and woe.