Recover from complex numbers with multiple parts
This first change properly deals with the situation when a number contains
multiple complex parts and/or multiple exponents.

Extra complex number parts (imaginary and arguments) are now properly
detected and reported as such. Previously we expected only one such part,
now we handle unlimited number. As ComplexScanningMode::Imaginary is now
an inadequate description, it is renamed to Subsequent. It is still
slightly different from Argument mode in how 'i's are handled.

Also we implement scanning of multiple consecutive exponents. The
exponents can have an optional sign in them. We now treat this sign
as a part of an extra exponent that needs to be scanned over during
recovery, not treated as a start of a new complex part. Care must be taken
to handle exponents in every place where we are panic-scanning (such as
infnan suffixes, for example).