Support for datum labels
This is simple as well. R7RS syntax is a bit ambiguous about whether these
tokens must be followed by an explicit delimiter, but the examples and the
wording in grammar suggest that they are not required. So we successfully
parse "#1##2#" as LabelRef("1"), LabelRef("2").

Some of the previous tests had to be updated as they actually include
invalid LabelMark tokens.