Support for directives
This only scans over them and handles errors. There are no side effects
on the identifier scanning now.

As you can see, directives do produce tangible tokens. This should be
easier to handle for external tools that have to filter out whitespace
and comments anyway, and it allows us to fully reconstruct the character
stream out of the token stream.