Case-folding support for ASCII
Initially I thought that character names are always case-sensitive, with
this assumption supported by items 6.6 and 7.1 of R7RS:

  Case is significant in #\<character> and in #<character name>,
  but not in #\<hex scalar value>.

  Case is not significant except in the definitions of <letter>,
  <character name> and <mnemonic escape>; for example, #x1A and
  #X1a are equivalent, but foo and Foo and #\space and #\Space
  are distinct.

However, here is how case-folding directives are described in item 2.1:

  The #!fold-case directive causes subsequent identifiers and
  character names to be case-folded as if by string-foldcase
  (see section 6.7). It has no effect on character literals.

Thus I assume that #!fold-case affects the case-folding of character
names as well.

Another somewhat important point is the case-folding of escaped
identifiers. R7RS does not specify this directly. Even more, it can
be read as implicitly assuming that escaped identifiers should be
case-folded as well:

  For example, the identifier |H\x65;llo| is the same identifier
  as Hello, <...>

But I believe that the case-folding behavior was included for
compatibility reasons and the escaped identifer form is explicitly
designed for writing verbatim symbols. Furthermore, I plan to apply
NFKC to non-verbatim identifiers in the future. Thus it makes sense
to case-fold only non-verbatim identifiers while leaving the verbatim
ones as is.