Unicode normalization
This is the normalization code I originally wrote for Sash. It has been
thoroughly tested and optimized so I believe in it.

The only major change is in the generator script in Python. I cleaned it
up, optimized the computation times, and switched over to using JSON
instead of pickle for data cache as JSON loads much faster.