Unicode case folding
This adds case-folding functions to the Unicode library. Currently we need
only toNFKC_Casefold() transformation, so we implement only it.

I do not know what data format will be the best for case-folding tables.
This needs some additional research and a look at other Unicode
implementations, preferably the ones using UTF-8. The usual tries are used
here, but maybe they are suboptimal. Though, I believe this is the best we
can do for NFKC_Casefold mapping. Maybe other mappings could be reasonably
split into 1-to-1 simple mapping and a bunch of irregular mappings.