Simplify bytevector recovery
Stop treating bytevector as special 'hashed' token. Fall back to scanning
the token as a number with invalid prefix if we are not on a happy path.

Fixup the tests accordingly:

- read #u( as invalid identifier "#u" followed by a paren
- read #u16 and #U571 as integers with invalid prefix
- read #u90_ex.a@mp!le1 as a polar complex number with exponential
  parts and an invalid prefix
- read #8 as a label mark

Also drop some diagnostics and scan_unrecognized() which are no longer
needed. This achieves the goal of not using Token::Unrecognized for
anything but the most grave cases.