Relicense to dual MIT/Apache-2.0 license
As the current sole owner of the project I would like to relicense
it under the dual license that is generally used in the Rust world.
Generally, I'm happy with just MIT, but having patent protection
from Apache is fine by me as well, and it will be easier to inter-
operate with third-party Rust code, so why not.

I would also like to be more friendly to *gasp* possible future
contributors, so the contribution terms are now spelled out in a way
that does not require any extra bureaucratic shit. I am not a lawyer,
but I am open for a constructive dialogue on the level I can understand.

- Add LICENSE.Apache and LICENSE.MIT files with actual legalese.
- Update the LICENSE file to clarify the policy of the project.
- Update the boilerplate comments all over the code to include
  the new licensing terms and depersonalize the copyright claims.
- Add a "license" property to the crates.