Move test utilities to a separate crate
We are going to add a new intergration test for a parser soon, and it is
becoming a hassle to rebuild these utilities for each module once (and run
their tests each time). Thus the utilities are moved into their own crate
which is going to be built only once.

One note about dependencies. Note that 'utils' is a dev-dependency of
the 'libreader' crate and 'libreader' crate is a dependency of 'utils'.
There is no dependency loop here because the latter libreader is actually
'libreader with #[cfg(test)]', a separate crate. It's incompatible with
the usual libreader. These utilities are for tests only.