Improve equality comparisons in diffs
1. Relax the requirement for sequence diffs to be just PartialEq.

We do not have any strict requirement for reflexiveness of the comparison
for diffing so just use PartialEq instead of Eq.

2. Introduce ShallowPartialEq trait and use it for tree diffs.

It is unreasonable to require PartialEq implementations to be shallow.
They are most likely to be #[derived] and defining them as deep
comparisons is very useful for testing by itself (like, assert_eq!
should use deep comparison).

Thus introduce a new trait for shallow comparisons and reformulate
tree diffs using it instead of Eq.

3. Remove the 'flat_' prefix from tree diffs.

I do not think that we'll ever need a more complex diff algorithm for
these tests so drop the useless prefix. Also reshuffle the docs a bit.