Use unboxed closures as comparators for diffs
Previously we have been using actual closure trait objects because I suck
as Rust. Now we use idiomatic unboxed closures for better efficiency. The
actual type of comparator is now encoded into the type of diff() functions
which means that the optimizer can inline the closures as they have unique
distinct types.

Note that child_diff() accepts a reference to the closure. This is
required for recursion, but should not affect the performance in any way.