Overhaul TreeNode trait
Previously this trait has been defined as returning a vector of references
to child nodes and expected that the tree node itself is a value. While we
can live with the vector thing, but the value interpretation is getting in
the way of generic algorithm implementation. For example, pretty-printing
trees currently assumes that _the nodes_ are printable, not the values
that they carry. We need a new design of the tree node trait to rectify
this issue.

New TreeNode trait explicitly specifies the value of the node. It may be
defined to be Self, or the type of some field, or even something else.
While we're here I have also adopted returning an iterator from children()
method as I now suck less at Rust and can actually define this properly.

All current users of TreeNode (diff::tree and pretty_print) have been
updated to use the new interface. While changes in the diffing are obvious
and minimal, pretty-printing has lost is DisplayTreeNode trait. Now the
values of the Tree are required to implement Display for the tree to be
pretty-printable without an explicit value formatter. The formatter itself
now also works for values, not the nodes.

Pretty-printing is going to be reimplemented into a trait soon so I did
not bother with updating its docs.