Drop tree diff and pretty-printing support
I've spent around two weeks fruitlessly trying to write a common interface
that can be used for both diffing and pretty-printing tree-like things.
The goal was to get a good tooling for parser tests and avoid rewriting
stuff in future when we add more trees after parse trees.

However, it did not work well *but* did divert my attention from actually
doing shit that matters, like writing the actual parser instead of
tinkering with the 'generic' interfaces. It makes it obvious that I have
no clue what good interface is, so I'd like to just throw out this code
to not bother with this idea anymore. Not that the diffing code in the
lexer was very helpful either, to be frank.