Support for bytevectors
This adds parsing of bytevector literals. They are considered simple data
as they do not allow nesting of arbitrary expressions. Actually, they
allow only exact integer literals in [0; 255] range.

However, here we do not check whether the numbers fall into that range.
The numbers are stored as just atoms and their parsing and validation is
left for future. The reason for this is that Scheme has far too many forms
of numbers. For example, 500/5, #b#e1100100, 0.01e+4, #O144@0, +100-0i
all represent effectively the same number 100 which is obviously a valid
bytevector constituent.

We also do handle nested data (currently allowing only nested
bytevectors) in order to check for proper bracket usage.

In order to simplify further implementation of vectors the parser is
bestowed with a lookahead of 1 (one) token in the form of `self.cur`
field. This makes the parser similar to the scanner.