Support for dotted lists
This adds support for dotted lists. We handle incorrect forms of dotted
lists by simply reporting all dots and falling back to treating the list
as a proper one.

Note that DatumValue::DottedList still contains just a Vec. It is
guaranteed that that vector contains at least two elements and the last
one is the cdr of the dotted list.

Also we remove some code duplication in datum test builders, and teach
them to accept IntoIterators rather than Vecs. This allows that fancy
implementation of line_sequence :)