Support for labels
This implements parsing of labels. Label references are similar to
simple data. Label marks are similar to abbreviations.

Note that the parser does not resolve nor check circular references.
This will done at some later stage. Currently we only accurately
represent the labeled locations and references to them.

As for the changes, one diagnostic was renamed to be more inclusive.
I think it will be used for datum comments as well because the error
is similar to what we have in abbreviations and labels.

One final thing that I would like to note is that only now is the
moment when I realised that the grammar maybe allows labels to have
spaces between their parts: "# 0 = ( 1 . # 0 # )" ought to be a
valid form of a circular list.

Consider the following:

    7.1.2. External representations

    <datum> --> <simple datum> | <compound datum>
             |  <label> = <datum>
	     |  <label> #

    <label> --> # <uinteger 10>

    7.1.1. Lexical structure

    Intertoken space can occur on either side of any token,
    but not within a token.

If "#" and "=" are considered 'tokens' then the above string should be
valid. However, if only <token> is considered a token that can be
followed by intertoken space then it's fine.

Racket and SBCL do not allow this interspaced form, so I guess it's fine
to leave it as it is now (not allowing spaces). But to be honest I would
have left it alone anyway. Treating the lone hash and an identifier =
specially in some cases will be too bothersome.