Handle unmatched closing parentheses
This is the last unimplemented!() in the parser code. Handle this case
exactly as extra dots. The diagnostic is non-fatal because we can
recover from this condition by simply ignoring the extra parentheses.

Besides, this case is very unlikely to occur in practice as Scheme
programs are heavily nested and are usually edited with proper editors
which can display matching parentheses. So the user has to try extra
hard to make this mistake.