Basic macro expander
This is the basic macro expander. It can handle all literals as well
as all basic special forms of Scheme. It does not handle user-defined
macros or modules, but R4RS should be fine.

As you may see by comparing the author date and the commit date, it's
pretty dated (and I'm pretty slacky). Initially I intended to split this
into subcommits like I did with the parser and lexer, but it turned out
to be too much of a hassle--for almost a year! So here you have one big
turd of the code.

The core idea here is the ExpanderStack. It allows combining existing
expanders and adding new ones as new macros are defined. The module
handling code will create the initial macro expansion environment and
then various define-syntax and let-syntax forms will expand it by
pushing new expanders onto the stack.

I won't explain the implementation details here too much because I'm
lazy. The code is pretty well commented so the workings should be
obvious. These are pretty straightforward transformations.

In the future this code may be moved into a separate libexpand library,
but currently it will stay here I guess. Until I implement some basic
interpreter at least.