Basic meaning module
This is the 'pretreatment' stage as described in 'Lisp in Small Pieces'
book. It is the semantic analysis pass on the expressions generated by
macroexpansion. Here we are going to resolve and check all variable
reference and covert the expressions into slightly more direct
intermediate representation, laying foundation for the middle-end.
Basically, this is where our front-end ends.