Make Environment a struct
It turned out that all environments have similar structure: an indexed
list of variables and a parent environment. So instead of using traits
let's just encode all environments with a single struct + 'kind' enum.

In the future we would like to store, share, and reuse environments
(e.g., to keep the local environment in closures for debugging),
so make the parent environments ref-counted (instead of using an
Option<&'a Environment<'a>>). This will greatly simplify the code if
the environments are stored in Meaning structs (don't need to pull
lifetimes into that).

Environments are immutable after they are created, so we simply use Rc.
We do not need RefCell or anything like that.