Treating variable errors
Currently there are two kinds of errors which are immediately detectable
during the semantic analysis:

- references to undefined variables
- assignment to immutable variables

(There will be more later.)

In order to report these diagnostics we will need to pass along a
reference to Handler which is used by other code. Update the signatures
for that.

We will also need to recover from these errors. For references we will
return a special placeholder to indicate a faulty expression. Invalid
assignments will be replaced by computation of the new variable value in
order to preserve any side-effects in that computation for later
analysis.

Note that the Handler always expects a Span while Variable stores only
Option<Span>. While variables can have no location if they are
introduced by macroexpansion, the variables which can cause the errors
described above must be introduced by the user in the source code, and
thus will have a Span. The compiler should never introduce a variable
that cannot be resolved.