Splicing (begin ...) forms
One particular wart of Scheme is the 'splicing' (begin ...) form defined
by the item 4.2.3 "Sequencing" of R7RS. A rant about it is included in
the comments for anyone interested. But we will still support it because
of backwards compatibility.

The custom iterator uses std::slice::Iter directly because I was not
able to abstract it out with traits. Maybe I'll get better at Rust later
and will do something about it. But I guess I'm fine with it as it is.

Also add some more tests to verify that 'begin' is spliced in only where
it is supposed to be: at the top-level and in bodies (ultimately defined
by lambda forms). It is not spliced in expression contexts.