Represent constants as indices
We are moving away from inlining the constants into the meaning of the
programs. The constants are now gathered into a separate vector during
processing, and their representation is now a simple index into that
vector.

We do not perform any deduplication at this step. Deduplication is not
essential at the current moment and it's an optimization anyway.

The tests for constants are also really lacking at the moment. We'll get
back to them after some more readability improvements.

Note that we also do not support whole range of Scheme values yet. This
will be improved later, I suppose. We'll need to be careful with all
'value' representations we have: ScannedDatum, Literal, Value. For
example, Scheme semantics require the empty list to be a singleton which
is not accurately conveyed by ScannedDatum and Literal. Furthermore,
this should be the time and place where we get rid of the InternPool,
parse the numbers, and resolve the datum labels.