Verify constant meanings
Introduce a way to verify that meaning() produces correct constant list.
It is done in kinda roundabout way with a closure because currently
Values require an InternPool for their construction, so we really need
to have access to the pool used by the test.

Note that if the constant list is not specified then it is not checked
by the test. This is because we're not really interested in constants in
every single test. They are collected in pretty simple way, so it's
enough to verify them in only some of the tests.