Run tests under kcov, post results to coveralls.io
We *could* have installed 'cargo-kcov' package, but its too 'trendy and
modern'. Read: it uses a shitload of 3rdparty dependencies. One of those
dependencies is 'clap' which fails to build on a machine with 512 GB of
RAM as it's too low. Which is ridiculous and fails our build in turn.

That's why we invoke kcov manually. We ask cargo to only build the
tests, but not run them. After that we manually scan for all executable
files in the build directory and run them all. This process does not run
doctests, but we'll be fine with that for now.

Note that kcov is run with a '--verify' option. This is essential
because the machine code generated by rustc make kcov puke and segfault
when breakpoints are inserted into it.

Also note that all coverage reports are put into a single directory.
This is essential for kcov to posts them all together to coveralls.io.
Without this only the last run will be posted.

Finally note the explicit --strip-path option. We need it to get correct
relative paths which allow coveralls to display source code in reports.