Split intern pool into submodules
We're going to beef up the Atom implementation soon so it will be a good
idea to split up the interning into submodules. We can move atoms and
interned strings into their own modules and keep the pool implementation
in the main module.

Note that we keep the exports the same for compatibility. Another thing
worth noting is the "pub(super)" exporting of the Atom implementation.
I did not know that Rust allows that, but as you can see it's quite
useful for the case when a submodule is used as an implementation
detail.