Custom Debug impl for Atom
This provides a custom implementation of debugging format "{:?}" for
atoms. Instead of rendering atoms as "Atom(123)" we now allow to specify
an intern pool which will be used to retrieve corresponding strings for
formatting.

One issue with that is that Debug trait does not allow any
configuration. We solve it by using dynamically-scoped variables
(aka fluid variables). These are a special kind of global thread-local
variables. They are commonly used in Lisp family of languages for
configuration.

Now you can control the formatting of atoms by specifiying the pool to
use in `with_format_pool()` function call. This temporarily sets the
current pool for the duration of the call and restores it back later.

I have tried making a generic interface to them, but Rust's interface to
thread-local variables makes it kinda hard. So I gave up for now and
hardcoded the accessors. Maybe sometime...

The implementation is a bit wonky, but I really wanted to get a feel of
using UnsafeCell. It's not strictly necessary here (RefCell will do),
but we'll still need unsafe code for casting the references, so why not,
we can go full ham on the unsafe and keep the runtime overhead minimal.