New testing harness for parser
I really liked the idea of using debugging formats for testing (see
tests for the `meaning()` function in libeval). Let's apply it to
previously written code as well.

Before that I though that it's really important to compare all fields in
results, including the spans. But now I realise that we need spans only
for diagnostic reporting. Other than that they have zero semantic value.
So tests should not really check for them (maybe with an exception of
the lexer tests). Spans should be meaningful only for diagnostics.

This commit implements a new test harness based on a builder pattern.
The comparison is now performed on the rendered representations of
expected and actual results. (Diagnostics are still compared as it).

We'll update the actual tests in the next commit.