New testing harness for expander
This is the same one as we use for parser now. In the next few commits
we will update the tests and drop the old harness.

Also note that the new harness accepts just &SchemeBase, not an
ExpanderFactory. I have noticed that there is only one implementation
of that trait, so we don't really needed it (and the generic code).