Use the same style for expected results
While I was working on new tests for parser and expander I have realized
that we do not really need to perform whitespace-ignoring comparison of
the actual and expected rendered forms in tests. We can use Rust's
string continuation syntax to keep the strings readable instead. Update
the tests for semantic analysis to use the same formatting as expander
and parser tests, and drop the trim_space() function.