Move diagnostic reporting to arguments
We do not really need to store a reference to the diagnostic handler in
the expanders. It is needed only during the call to expand() method.
Passing the handler via arguments allows it to have a different lifetime
than the expander, removing all those pesky lifetimes from everywhere.
This in turn will allow storing the expanders on the heap (e.g., in an
Rc) which will be needed to keep them in environments.