Make diagnostic spans mandatory
It turned out that most of the time we report diagnostics with known
spans.  Moreover, it seems logical than any diagnostic that we want to
report to the user should be associated with some source code. Anything
else is compiler warning, error message, whatever, but not a diagnostic
which is intrinsically associated with the source code.

Remove the Option from diagnostics' spans. Also rename the field to
"span" as it is the name we use in other places together with "kind".