Make meaning spans mandatory
The same as expressions, their meanings are also associated with user
input so they should always have spans. The only edge case is when we
encounter empty expression sequence (can be a result of an empty file).
We need to use some span in this case. The ideal one would be a span of
the whole file, but we don't have it because the parser produced no
tokens, no s-expressions, no nothing. We have no idea of the span.
Well... it's not like we're going to report any diagnostics there or
emit any code, so we can use some bogus value and get done with that.