Return definition spans of resolved variables
We now store the definition spans of the variables in environments, but
they are not returned back when the variables are resolved. Return them.

For unresolved variables we have to invent some span. The caller should
disregard the returned definition span. I don't like this so it will be
encoded into the type system in the next commit.