Replace ReferenceKind::Unresolved with Option
This reference kind kinda broken the type safety of the references.
It could be returned by environments as a result of variable resoltion,
but the returned definition span contained bogus value.

Instead of this ugly hack we can use a better idiomatic approach with
Option. Return Option<VariableReference> from environments, use None for
unresolved variables.

Assignment looks kinda ugly with that, but that's because we want to
produce diagnostics nicely ordered. Maybe I will find a better way...