Store syntactic variables in environments
This adds an ability to store syntax transformers in environments.
They are stored as Box<Expander>, with a new Expander trait. This time
we require the expanders to always produce _some_ expression, even in
case of errors. Recovery should be possible, so it's okay to only report
errors to the diagnostic handler.

VariableReferences now can carry references to the expanders, so they
need an explicit lifetime (to tell the fact that the stored reference
lives no longer than VariableReference).

New enum variants mean that we need to handle them in existing code.
We add some dummy handlers for now. We'll get back to the TODOs later.