New core expander
This implements a new expander entry point. It codifies the behavior of
BasicExpander and ApplicationExpander as this is really the kernel
expander behavior of Scheme. Literals always mean literals, forms always
mean applications (when they are not macro uses), identifiers always
mean variable references (as Scheme does not support identifier macros).
Thus we can hardocode this behavior into the core expander itself rather
than abstract it out in some BasicExpander implementing the Expander
trait, which must be the lowest expander in the stack to work properly.