New expander for (begin expr1 expr2 ...)
Implement the new expander trait (named Expander) for BeginExpander.
Rename the old expander into OldExpander to avoid name conflicts. The
old trait is still used as it is still imported as Expander elsewhere.

Note the expect_form() utility function. It panics if applied to
something that is not a valid form, or is an unexpected form. This
requirement is currently enforced by the core expander, which calls
the expanders only for intended forms. I would like it to stay this way.