Use new expander in expander tests
Update the code that constructs the test harness to use the new
expander. This will require properly constructed environments, so
update the environment code as well. The SchemeBase struct is now
a dumb storage for the names, so rename it, the fields, and the
accessors appropriately.

Also update the tests for quotations. Now they yield the first term
instead of the last one if there are more than one term in the form.
Consistency with begin was not a too strong argument for that.