Use Undefined for incomplete "begin" forms
Empty (begin) expression has undefined meaning in Scheme. Do not defer
this to semantic analysis. It's okay because _the syntax_ of "begin"
form requires that it has at least one term.

For some weird reason rustc can't infer the type of "expressions" to be
Vec<T>, because it does not see that it is used only in one arm of the
conditional. We have to specify the type explicitly via a turbofish.