Common term count checking
The new expaders are all checking whether they are passed a correct
number of terms. This code is almost identical in all places, so
provide a common implementation for that. It could be used later for
user-defined macros as well.

The new utility uses the Into trait to have a convenient interface
with argument overloads (this is how you do it in Rust). You can
specify the bounds on the expected term counts with either a single
value, or a range.

When writing this function I have finally realized my intentions behind
these obscure 'terms.len() >= 2' lines in begin and quote when reporting
dotted forms. I wanted the reports about dots to be suppressed if the
dots fall into the range already reported as extra forms. Thus the new
utility contains a common check for this. It also updates the rang for
the extra forms a bit.

Here's how a report looked before the change:

    (if #f 1 2 3 . 4)
               ^---~
	        ^-~

and here's how it should look now:

    (if #f 1 2 3 . 4)
              ^----~

(The entire extra span is marked.)