Fix a typo in licensing statement
This has been bugging me for a while. Now I guess is a great time
to fix it! Also remove the "contributing" section from README. It's
kind of self-evident (and can be inferred from the license text).

I also pondered the prospect of relicensing all of the code simply
under Apache license, but that's too much of a bother to update all
of the copyright lines in the codebase.