Update Unicode table tool
Mark it executable (no idea why it wasn't) and use FTP to access UCD
instead of HTTP. For some weird reason unicode.org likes to dish out
month-long IP subnet bans if you try downloading more than a couple
of megabytes of the database via HTTP. Or that's just my ISP being
shitty as it is. Whatever...

(As a side note: the tool still works fine and I'll probably continue
using it as is. We still have a pending upgrade from 9.0.0 to 12.0.0.)