New diagnostic reporting interface
Introduce a new DiagnosticBuilder object that can be used to report
the diagnostics to the specified Handler with a delay. Initially I
wanted to use global dynamic variables (like for Atom formatting),
however that turned out to be a bad idea in practice. Global state
does not allow any concurrent usage of multiple parsers which we may
need in the future.

Now the diagnostics are constructed directly from DiagnosticKind,
and they are automatically reported after that. (Ab)using Drop for
this kind of job is arguably not good but it makes a really nice
fluid interface so this should be fine.

This is mostly cosmetic change, but it plays with rustfmt better than
the current interface. Having nice rustfmt is one of the goals behind
this change. The other one is really having some builder for partially
constructed diagnostics: later we might need to augment diagnostics
with notes and stuff.

While we're here, do impl From<Range> for Span. This gives spans
a convenient constructor, and we can now use "impl Into<Span>" trait
bounds for _really_ convenient arguments. After two years of doing Rust
I can finally do this :)